Update entrypoint.sh
Added a check to verify is the .sh is starting with a user set by docker, if it is, it skip the permission setup and gosu. If not, it does it's old behaviors.

Also re-introduced the PORT environment variable. Not everyone will run that program on docker, and conflict may arise with others program, putting aside the fact that 8080 is commonly used, begin able to change the PORT without the needed of recompiling everything is useful.